Move help.supportLevel state to happychat.user
