+++ client/state/help/reducer.js
@@ -1,9 +1,5 @@
 import { withStorageKey } from '@automattic/state-utils';
-import {
-	HELP_CONTACT_FORM_SITE_SELECT,
-	SUPPORT_HISTORY_SET,
-	SUPPORT_LEVEL_SET,
-} from 'calypso/state/action-types';
+import { HELP_CONTACT_FORM_SITE_SELECT, SUPPORT_HISTORY_SET } from 'calypso/state/action-types';
 import { combineReducers } from 'calypso/state/utils';
 import courses from './courses/reducer';
 import directly from './directly/reducer';
@@ -43,31 +39,12 @@ export const supportHistory = ( state = [], { type, items } ) => {
 	}
 };
 
-/**
- * The level of support we're offering to this user (represents their highest paid plan).
- *
- * @param {object} state  Current state
- * @param {object} action Action payload
- * @param action.type
- * @param action.level
- * @returns {object}        Updated state
- */
-export const supportLevel = ( state = null, { type, level } ) => {
-	switch ( type ) {
-		case SUPPORT_LEVEL_SET:
-			return level;
-		default:
-			return state;
-	}
-};
-
 const combinedReducer = combineReducers( {
 	courses,
 	directly,
 	ticket,
 	selectedSiteId,
 	supportHistory,
-	supportLevel,
 } );
 
 export default withStorageKey( 'help', combinedReducer );
